+++ kotlinx-coroutines-debug/test/DebugTestBase.kt
@@ -16,7 +16,7 @@ open class DebugTestBase : TestBase() {
     open fun setUp() {
         before()
         DebugProbes.sanitizeStackTraces = false
-        DebugProbes.enableCreationStackTraces = true
+        DebugProbes.enableCreationStackTraces = false
         DebugProbes.install()
     }
